remove now unused oberver service
